#415: Lager spesifikke spørringar der dei høyrer heime i staden for å eksponere entitymanager rundt i heile basen

+++ src/main/kotlin/no/roedt/person/PersonRepository.kt
@@ -69,5 +69,5 @@ class PersonRepository : PanacheRepositoryBase<Person, Int> {
 
     fun hypersysIDTilRingerId(userId: UserId) = entityManager.list(
         "select ringer.id from ringer inner join person on person.id = ringer.personId and person.hypersysID = ${userId.userId} "
-    ).first()
+    ).first() as Int
 }
